Support a mac build system. Fix overall warnings. #3
Adds static_assert for struct size assertion when available.
Update all platform-dependent files to use platform.h
Replace STDCALL with SPHERE_STDCALL to remove hundreds of warnings.
Add SConscripts for engine/mac and video_mac/*.
Add SPHERE_HAS_NETWORKING to platforms.h and disable it for Mac build.
Enable both the unix and mac build, for mac, to compile and work (not) the same.
Use stdint in types.h so that all types are the intended size on every system:
 fixes assertion errors on packed structs.
Fix a lot of warnings from Clang.
Switch to C++11 mode when on Mac.

Warranty: build is only tested on Mac. A rundown on linux and windows should be
done to make sure MVC does not freak out about real C++ stuff and that SCons still
works.
Current status of the Mac build: executing but failing with corona, locally.